# be/internal/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Minh Quang Personnel API
  version: 1.0.0
  description: API quản lý tài khoản quản trị và Huynh đệ công quả tạm thời.
servers:
  - url: http://localhost:8080
tags:
  - name: Authentication
  - name: Users
  - name: Volunteers
  - name: Departments
paths:
  /healthz:
    get:
      summary: Kiểm tra trạng thái API
      responses:
        "200":
          description: API hoạt động
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Đăng nhập
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        "200":
          description: Đăng nhập thành công
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/me:
    get:
      tags: [Authentication]
      summary: Lấy tài khoản hiện tại
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Tài khoản hiện tại
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/logout:
    post:
      tags: [Authentication]
      summary: Đăng xuất
      security: [{ bearerAuth: [] }]
      responses:
        "204":
          description: Đã đăng xuất
  /api/users:
    get:
      tags: [Users]
      summary: Danh sách tài khoản quản trị
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Danh sách tài khoản
    post:
      tags: [Users]
      summary: Tạo tài khoản quản trị
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, display_name, password]
              properties:
                username: { type: string }
                display_name: { type: string }
                password: { type: string, minLength: 8 }
      responses:
        "201":
          description: Tài khoản đã tạo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /api/volunteers:
    get:
      tags: [Volunteers]
      summary: Danh sách hồ sơ công quả
      security: [{ bearerAuth: [] }]
      parameters:
        - name: q
          in: query
          schema: { type: string }
          description: Tìm theo họ tên, pháp danh, số điện thoại hoặc nơi sinh hoạt.
        - name: status
          in: query
          schema: { type: string, enum: [active, departed] }
      responses:
        "200":
          description: Danh sách hồ sơ
    post:
      tags: [Volunteers]
      summary: Thêm hồ sơ công quả
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VolunteerInput"
      responses:
        "201":
          description: Hồ sơ đã tạo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Volunteer"
  /api/volunteers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Volunteers]
      summary: Xem hồ sơ
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Chi tiết hồ sơ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Volunteer"
    put:
      tags: [Volunteers]
      summary: Cập nhật hồ sơ
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VolunteerInput"
      responses:
        "200":
          description: Hồ sơ đã cập nhật
    delete:
      tags: [Volunteers]
      summary: Xoá hồ sơ
      security: [{ bearerAuth: [] }]
      responses:
        "204":
          description: Đã xoá hồ sơ
  /api/volunteer-options/departments:
    get:
      tags: [Volunteers]
      summary: Gợi ý các phân ban đã sử dụng
      security: [{ bearerAuth: [] }]
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 20, default: 10 }
      responses:
        "200":
          description: Danh sách phân ban đang sử dụng
  /api/departments:
    get:
      tags: [Departments]
      summary: Danh sách phân ban
      security: [{ bearerAuth: [] }]
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: active
          in: query
          schema: { type: string, enum: [all, "true", "false"], default: all }
      responses:
        "200":
          description: Danh sách phân ban và số hồ sơ đang sử dụng
    post:
      tags: [Departments]
      summary: Thêm phân ban
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 60 }
      responses:
        "201":
          description: Phân ban đã tạo
  /api/departments/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    put:
      tags: [Departments]
      summary: Đổi tên phân ban
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Phân ban đã cập nhật }
    delete:
      tags: [Departments]
      summary: Xoá phân ban chưa có hồ sơ sử dụng
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Đã xoá phân ban }
        "409": { description: Phân ban đang có hồ sơ sử dụng }
  /api/departments/{id}/status:
    patch:
      tags: [Departments]
      summary: Bật hoặc ngừng sử dụng phân ban
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Trạng thái đã cập nhật }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    Unauthorized:
      description: Chưa đăng nhập hoặc phiên đăng nhập hết hạn
  schemas:
    User:
      type: object
      required: [id, username, display_name, role, active, created_at, updated_at]
      properties:
        id: { type: string }
        username: { type: string }
        display_name: { type: string }
        role: { type: string, enum: [admin] }
        active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    LoginResult:
      type: object
      required: [token, expires_at, user]
      properties:
        token: { type: string }
        expires_at: { type: string, format: date-time }
        user: { $ref: "#/components/schemas/User" }
    VolunteerInput:
      type: object
      required: [full_name, arrival_date]
      properties:
        full_name: { type: string }
        dharma_name: { type: string }
        birth_date:
          type: string
          description: Ngày hoặc năm sinh dạng văn bản tự do.
        cultivation_place: { type: string }
        phone: { type: string }
        department: { type: string, maxLength: 60 }
        notes: { type: string }
        avatar_url: { type: string, format: uri }
        arrival_date: { type: string, format: date }
        departure_date: { type: string, format: date }
    Volunteer:
      allOf:
        - $ref: "#/components/schemas/VolunteerInput"
        - type: object
          required: [id, status, created_at, updated_at]
          properties:
            id: { type: string }
            department_id: { type: string }
            status: { type: string, enum: [active, departed] }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    Department:
      type: object
      required: [id, name, active, volunteer_count, created_at, updated_at]
      properties:
        id: { type: string }
        name: { type: string, maxLength: 60 }
        active: { type: boolean }
        volunteer_count: { type: integer, minimum: 0 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }